+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.0] - 23/10/2021 - DD/MM/YYYY
+
+### Added
+
+- Grunt css build
+
+### Changed
+
+- Template markdown
+
 ## [0.0.6] - 30/08/2021 - DD/MM/YYYY
 
 ### Fix